move `signatures` logic into `canonicalize-candidates`
We only really need it in this file, this will also make it a bit easier
to further enhance the design system storage.
